refactor(rds): split service.rs by resource family (4245 -> 912 lines)
Audit 2026-05-19 P4 (partial). 44-method service.rs split into 10
family submodules under service/:

- instances.rs         1085  (create/delete/modify/reboot)
- mod.rs                912  (AwsService dispatch + lifecycle)
- snapshots.rs          505
- cluster_snapshots.rs  420
- restore.rs            351
- parameter_groups.rs   331
- subnet_groups.rs      231
- replicas.rs           193
- log_files.rs          186
- tags.rs                92
- engine.rs              42

extras.rs (4795 lines) intentionally left untouched: its bulk is a
single 1265-line macro-driven `handle_extra_action` dispatch fn
plus XML/parse helpers and tests — splitting by resource family
doesn't help that shape. Plan separately as a concerns-based split.

Same pattern as PR #1479 / #1480 / #1481. 205 unit tests pass.
